chore(mypy): wave 2 — lock 8 more modules in CI
Adds: contracts (rest beyond v4/telemetry), config, ports, evaluation,
context, metrics, planning, causality.

Combined wave 1+2 surface: 12 modules / 86 source files at 0 mypy errors.

Direct fixes (17 errors → 0):
- contracts/envelopes/agent_envelope.py: pass name/metadata explicitly to
  ParticipantRef constructor (4)
- evaluation/scoring.py: max(..., key=) lambda instead of bound method (1)
- context/detector.py: dataclasses.field(default_factory=list) for
  detected_keywords (1)
- context/composer.py: cast .get("response","") to str (1)
- metrics/emotional_kpi_engine.py: cast yaml.safe_load and float()
  np.mean / np.dot results (3)
- planning/goal_persistence.py: explicit type hints on conflicts list,
  by_status dict, to_dict result (3)
- causality/root_cause.py: float() wrap on min() result (1)
- causality/simulator.py: filter None for state_after, rename val to
  node_val for narrowing, type comparison dict[str, Any] (3)

+++ phionyx_core/causality/root_cause.py
@@ -269,7 +269,7 @@ class RootCauseAnalyzer:
             std = (sum((v - avg) ** 2 for v in values) / len(values)) ** 0.5
             if std > 0:
                 z_score = abs(node.current_value - avg) / std
-                return min(1.0, z_score / 3.0)  # Normalize: 3σ = 1.0
+                return float(min(1.0, z_score / 3.0))  # Normalize: 3σ = 1.0
 
         # Fallback: check against expected range
         return self._anomaly_magnitude(node.current_value, expected_range)
